refactor: moved sanity models
