+++ app/[locale]/participant/[participantId]/page.tsx
@@ -6,8 +6,8 @@ import { getClient } from "@/lib/api/graphql-client";
 import { gql } from "@/lib/graphql";
 import { GetInvitationQuery, InvitationDto } from "@/lib/graphql/graphql";
 import { logger } from "@/lib/logger";
-import { client } from "@/lib/sanity";
-import { ParticipantPageData } from "@/lib/types";
+import { client } from "@/lib/sanity/sanity";
+import { ParticipantPageData } from "@/lib/sanity/types";
 import { PortableText } from "next-sanity";
 import Image from "next/image";
 import styles from "./page.module.css";
