+++ app/components/OurStory.tsx
@@ -1,7 +1,7 @@
-import { TimelineItem } from '../../lib/types';
-import Image from 'next/image';
-import { urlFor } from '../../lib/sanity';
-import styles from './OurStory.module.css';
+import { TimelineItem } from "../../lib/sanity/types";
+import Image from "next/image";
+import { urlFor } from "../../lib/sanity/sanity";
+import styles from "./OurStory.module.css";
 
 interface OurStoryProps {
   data: TimelineItem[];
@@ -13,28 +13,31 @@ export default function OurStory({ data }: OurStoryProps) {
   return (
     <section className={styles.storySection} id="story">
       {data.map((moment: TimelineItem, index: number) => {
-        const isRight = moment.imageSide === 'right';
+        const isRight = moment.imageSide === "right";
         return (
-            <div key={index} className={`${styles.moment} ${isRight ? styles.right : styles.left}`}>
-                <div className={styles.imageCol}>
-                    {moment.image && (
-                        <div className={styles.imageWrapper}>
-                             <Image
-                                src={urlFor(moment.image).width(800).url()}
-                                alt={moment.title}
-                                width={400}
-                                height={483}
-                                className={styles.momentImage}
-                             />
-                        </div>
-                    )}
+          <div
+            key={index}
+            className={`${styles.moment} ${isRight ? styles.right : styles.left}`}
+          >
+            <div className={styles.imageCol}>
+              {moment.image && (
+                <div className={styles.imageWrapper}>
+                  <Image
+                    src={urlFor(moment.image).width(800).url()}
+                    alt={moment.title}
+                    width={400}
+                    height={483}
+                    className={styles.momentImage}
+                  />
                 </div>
-                <div className={styles.textCol}>
-                    <h2 className={styles.momentTitle}>{moment.title}</h2>
-                    <p className={styles.momentDesc}>{moment.description}</p>
-                </div>
+              )}
             </div>
-        )
+            <div className={styles.textCol}>
+              <h2 className={styles.momentTitle}>{moment.title}</h2>
+              <p className={styles.momentDesc}>{moment.description}</p>
+            </div>
+          </div>
+        );
       })}
     </section>
   );
